Create separate component to view and zoom in on image
This enables us to move the state and event handlers
out of the [[...id]] page component
